data/flows/news_etl/pyproject.toml
[project]
name = "news_etl"
version = "0.1.0"
description = "News ETL package"
readme = "README.md"
requires-python = "==3.13.1"
dependencies = [
    "common",
    "newspaper4k>=0.9.0",
    "sentence-transformers>=3.0.0",
]

[tool.uv.sources]
common = { workspace = true }
